Revamped tests
