+++ src/__tests__/test_chap1.ts
@@ -1,7 +1,6 @@
-import { grader, student } from './examples/chap1'
 import { awsEventFactory } from './helpers'
-import { runAll } from '../index'
-import 'jest'
+import { runAll, TestCase } from '../index'
+import { Student, Grader } from './types';
 
 const makeAwsEvent = awsEventFactory({
   chapter: 1,
@@ -11,6 +10,94 @@ const makeAwsEvent = awsEventFactory({
   },
   globals: []
 })
+
+/*
+  Simple fibonacci function test.
+
+  No prepend.
+
+  Student program: correct, partial, wrong, syntax, runtime, timeout.
+
+  No postpend.
+
+  Testcase program: valid, syntax, runtime, timeout.
+*/
+
+const validStudentCorrect =
+  `const f = i => i === 0 ? 0 : i < 3 ? 1 : f(i-1) + f(i-2);`
+
+const validStudentWrong =
+  `const f = i => i < 3 ? 2 : f(i-2) + f(i-3);`
+
+const validStudentPartial =
+  `const f = i => i < 3 ? 1 : f(i-1) + f(i-2);`
+
+const invalidStudentRuntime =
+  `const f = i => f(j+1);`
+
+const invalidStudentSyntax =
+  `const f = i => i === 0 ? 0 : i < 3 ? 1 : f(i-1) + f(i-2)`
+
+// Does not compute fast enough to exceed max call stacks, relies on timeout
+const invalidStudentTimeout =
+  `const f = i => i < -3 ? 0 : f(i+1) + f(i+2);`
+
+const student: Student = {
+  valid: {
+    correct: validStudentCorrect,
+    wrong: validStudentWrong,
+    partial: validStudentPartial
+  },
+  invalid: {
+    runtime: invalidStudentRuntime,
+    syntax: invalidStudentSyntax,
+    timeout: invalidStudentTimeout
+  }
+}
+
+const validTestcases: TestCase[] = [
+  {
+    program: `f(1);`,
+    answer: "1",
+    score: 1
+  },
+  {
+    program: `f(3);`,
+    answer: "2",
+    score: 1
+  },
+  {
+    program: `f(5);`,
+    answer: "5",
+    score: 1
+  }
+]
+
+const invalidTestcaseSyntax = [
+  {
+    program: `f(1)`,
+    answer: "1",
+    score: 1
+  }
+]
+
+const invalidTestcaseRuntime = [
+  {
+    program: `g(1);`,
+    answer: "1",
+    score: 1
+  }
+]
+
+const grader: Grader = {
+  validPrepend: ``,
+  validPostpend: ``,
+  validTestcases: validTestcases,
+  invalidTestcases: {
+    runtime: invalidTestcaseRuntime,
+    syntax: invalidTestcaseSyntax
+  }
+}
 
 test('prepend OK, postpend OK, testCases OK, student OK, correct', async () => {
   const results = await runAll(makeAwsEvent({
